application/requirements.txt
kafka-python==2.0.2
tqdm==4.66.1